third_party/WebKit/LayoutTests/external/wpt/html/semantics/scripting-1/the-script-element/module/instantiation-error-5-expected.txt

This is a testharness.js-based test.
FAIL Test that loading a graph in which a module is already errored results an error. assert_array_equals: property 2, expected "Uncaught SyntaxError: The requested module './instantiation-error-5c.js' does not provide an export named 'something'" but got "Uncaught SyntaxError: The requested module './instantiation-error-5e.js' does not provide an export named 'something'"
Harness: the test ran to completion.
